Refactoring of my JS code
