+++ js/api/posts/create.mjs
@@ -1,14 +1,11 @@
 import { API_SOCIAL_URL } from "../constants.mjs";
-
 import { authFetch } from "../authFetch.mjs";
 
 const action = "/posts";
 const method = "post";
 
 export async function createPost(postData) {
-  // Legger til en forespørselparameter for å inkludere forfatterdata hvis APIet støtter det
   const createPostURL = `${API_SOCIAL_URL}${action}?_author=true`;
-
   const response = await authFetch(createPostURL, {
     method,
     headers: {
@@ -26,7 +23,7 @@ export async function createPost(postData) {
   }
 
   const data = await response.json();
-  console.log(data); // Logger data for å sjekke hva som returneres
+  console.log(data);
 
   return data;
 }
