+++ js/api/posts/delete.mjs
@@ -1,5 +1,4 @@
 import { API_SOCIAL_URL } from "../constants.mjs";
-
 import { authFetch } from "../authFetch.mjs";
 
 const action = "/posts";
@@ -7,7 +6,6 @@ const method = "delete";
 
 export async function removePost(id) {
   const removePostURL = `${API_SOCIAL_URL}${action}/${id}`;
-
   const response = await authFetch(removePostURL, {
     method,
   });
@@ -18,5 +16,5 @@ export async function removePost(id) {
     throw new Error("Failed to delete post: " + (await response.text()));
   }
 
-  return response.status; // Returnerer kun statuskoden ettersom det ikke er noe innhold
+  return response.status;
 }
