+++ js/api/posts/read.mjs
@@ -1,12 +1,10 @@
 import { API_SOCIAL_URL } from "../constants.mjs";
-
 import { authFetch } from "../authFetch.mjs";
 
 const action = "/posts";
 
 export async function getPosts() {
   const getPostsURL = `${API_SOCIAL_URL}${action}?_author=true`;
-
   const response = await authFetch(getPostsURL);
 
   return await response.json();
