+++ js/api/profiles/index.mjs
@@ -1,2 +1,3 @@
-export * from "./read.mjs";
+export * from "./getProfiles.mjs";
 export * from "./update.mjs";
+export * from "./getProfile.mjs";
