+++ js/components/hideLoader.mjs
@@ -1,0 +1,4 @@
+export function hideLoadingIndicator() {
+  const loadingIndicator = document.querySelector(".loading-indicator"); //
+  if (loadingIndicator) loadingIndicator.style.display = "none";
+}
