+++ js/components/index.mjs
@@ -1,4 +1,8 @@
 export * from "./openDeleteModal.mjs";
 export * from "./search.mjs";
 export * from "./sort.mjs";
-export * from "./loadingIndicator.mjs";
+export * from "./hideLoader.mjs";
+export * from "./showLoader.mjs";
+export * from "./hydrateEmailField.mjs";
+export * from "./getLoggedInUser.mjs";
+export * from "./countUserPosts.mjs";
