+++ js/components/validURL.mjs
@@ -1,0 +1,8 @@
+export function isValidURL(string) {
+  try {
+    new URL(string);
+    return true;
+  } catch (_) {
+    return false;
+  }
+}
