+++ js/handlers/createPost.mjs
@@ -1,9 +1,9 @@
 import { createPost } from "../api/posts/create.mjs";
-import { postTemplate } from "../templates/post.mjs"; // Sørger for riktig import
+import { postTemplate } from "../templates/post.mjs";
 
 export function setCreatePostFormListener() {
   const form = document.querySelector("#createPost");
-  const postsList = document.querySelector("#postsList"); // Sjekker at dette er ID-en for containeren hvor postene vises
+  const postsList = document.querySelector("#postsList");
 
   if (form) {
     form.addEventListener("submit", async (event) => {
@@ -26,8 +26,8 @@ export function setCreatePostFormListener() {
         const newPost = await createPost(postData);
         if (newPost && postsList) {
           const postElement = postTemplate(newPost);
-          postsList.insertBefore(postElement, postsList.firstChild); // Legger til den nye posten øverst i listen
-          form.reset(); // Tømmer skjemafeltene etter vellykket innsending
+          postsList.insertBefore(postElement, postsList.firstChild);
+          form.reset();
         }
       } catch (error) {
         console.error("Error creating post:", error);
