+++ js/handlers/index.mjs
@@ -4,5 +4,7 @@ export * from "./logOut.mjs";
 export * from "./register.mjs";
 export * from "./updatePost.mjs";
 export * from "./updateProfile.mjs";
-export * from "./deletePost.mjs";
 export * from "./displayPostCount.mjs";
+export * from "./searchListener.mjs";
+export * from "./sortListener.mjs";
+export * from "./updateSortDisplay.mjs";
