+++ js/handlers/login.mjs
@@ -1,16 +1,16 @@
-import { login } from "../api/auth/login.mjs"
+import { login } from "../api/auth/login.mjs";
 
 export function setLoginFormListener() {
   const form = document.querySelector("#loginForm");
 
   if (form) {
     form.addEventListener("submit", (event) => {
-      event.preventDefault()
+      event.preventDefault();
       const form = event.target;
       const formData = new FormData(form);
-      const profile = Object.fromEntries(formData.entries())
-  
-      login(profile)
-    })
+      const profile = Object.fromEntries(formData.entries());
+
+      login(profile);
+    });
   }
 }
