+++ js/templates/index.mjs
@@ -1,5 +1,6 @@
 export * from "./post.mjs";
 export * from "./posts.mjs";
+export * from "./renderPosts.mjs";
 export * from "./updateUserData.mjs";
 export * from "./renderUserPosts.mjs";
 export * from "./renderSinglePost.mjs";
