+++ js/templates/renderPosts.mjs
@@ -1,0 +1,5 @@
+import { postTemplate } from "./post.mjs";
+
+export async function renderPostTemplates(postDataList, parent) {
+  parent.append(...postDataList.map(postTemplate));
+}
